transport aware sessions refactored remote server logic out of index.ts
