+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@octomind/octomind-mcp",
-  "version": "1.1.2",
+  "version": "1.1.4",
   "description": "an mcp server for the octomind platform",
   "main": "./dist/index.js",
   "packageManager": "pnpm@10.7.1+sha512.2d92c86b7928dc8284f53494fb4201f983da65f0fb4f0d40baafa5cf628fa31dae3e5968f12466f17df7e97310e30f343a648baea1b9b350685dafafffdf5808",
