+++ src/version.ts
@@ -1,2 +1,2 @@
 // Generated by genversion.
-export const version = "1.1.2";
+export const version = "1.1.4";
